requirements.txt
moviepy==1.0.3
librosa==0.10.1
numpy<2.0.0
torch
torchvision
transformers
pillow
streamlit
ftfy
regex
tqdm
psutil
opencv-python
